+++ com.unity.netcode.gameobjects/Runtime/Messaging/Messages/ChangeOwnershipMessage.cs
@@ -332,8 +332,8 @@ namespace Unity.Netcode
             // Sanity check that we are not sending duplicated change ownership messages
             if (networkObject.OwnerClientId == OwnerClientId)
             {
-                UnityEngine.Debug.LogError($"Unnecessary ownership changed message for {NetworkObjectId}.");
-                // Ignore the message
+                // Log error and then ignore the message
+                UnityEngine.Debug.LogError($"Client-{context.SenderId} ({RequestClientId}) sent unnecessary ownership changed message for {NetworkObjectId}.");
                 return;
             }